New option limit_handler_unless_threshold
A new otpion 'limit_handler_unless_threshold' is created that works in a
similar way that 'limit_reporter_unless_threshold'.
When this option is enabled, users will only have access to issues which
are handled by themselves.
The option is a threshold that behaves in the same way than
'limit_reporter_unless_threshold'.

When there is a conflict in both limit_xxx options, the user will see
only those issues that are either reporter by him, or handled by him.

Fixes: #10831