Utility function to process event results
Encapsulate a common function to process the results of an event of
type "default". Usually, extracting the data from the event responses
requires a nested loop that is not simple.
This new function returns a simple array of the items extracted from the
event result.
An optional parameter for a callback function is provided to cover
complex scenarios.

This commit implements a refactor of two usages of this pattern. One is
a simple scenario for EVENT_MENU_MAIN/EVENT_MENU_MAIN_FRONT. The other
is EVENT_FILTER_COLUMNS impementing a callback for advanced processing
of the result.